refactor: move out optional attendees
- make component smaller and more readable
